requirements.txt
pre-commit==3.5.0
python-telegram-bot